+++ lib/dollar.dart
@@ -1,14 +1,5 @@
 library dollar;
 
-export 'src/core.dart'
-    show $handle, $effect, $if, $ref, $Ref, $Effect, $EffectHandler;
-export 'src/effects.dart'
-    show
-        $var,
-        $final,
-        $scan,
-        $listen,
-        $Var,
-        $UpdateVar,
-        $AddListener,
-        $RemoveListener;
+export 'src/core.dart';
+export 'src/effects.dart';
+export 'src/expressions.dart';