+++ lib/src/effects.dart
@@ -1,72 +1,4 @@
-import 'package:collection/collection.dart';
 import 'package:dollar/dollar.dart';
-
-$Var<T> $var<T>(T init()) {
-  final didInit = $ref(() => false);
-  final ref = $ref<$Var<T>>(() => null);
-  $if(!didInit.value, () {
-    ref.value = _$VarImpl(ref, init(), $effect);
-    didInit.value = true;
-  });
-  return ref.value;
-}
-
-T $final<T>(T init()) {
-  return $var(init).value;
-}
-
-T $scan<T>(T work(T prev), [Iterable keys]) {
-  final prevKeys = $ref<Iterable>(() => null);
-  final status = $ref<T>(() => null);
-
-  return $if(keys == null || !shallowEquals(prevKeys.value, keys), () {
-    status.value = work(status.value);
-    prevKeys.value = keys;
-    return status.value;
-  }, orElse: () => status.value);
-}
-
-R $listen<T, R>(R callback(T value)) {
-  final listener = $ref<R Function(T)>(() => callback);
-  listener.value = callback;
-  final result = $ref<R>(() => null);
-
-  $final(() {
-    $effect($AddListener(
-        $handle<T, R>((T event) => result.value = listener.value(event))));
-  });
-
-  return result.value;
-}
-
-abstract class $Var<T> {
-  T get value;
-  set value(T newValue);
-
-  T get() => value;
-  T set(T newValue) => value = newValue;
-}
-
-class _$VarImpl<T> extends $Var<T> {
-  @override
-  T get value => _value;
-
-  @override
-  set value(T newValue) {
-    final effect = $UpdateVar(_value, newValue, _ref);
-    _value = newValue;
-    _$effect(effect);
-  }
-
-  $Ref _ref;
-  T _value;
-  $EffectHandler _$effect;
-
-  _$VarImpl($Ref ref, T value, $EffectHandler $effect)
-      : _ref = ref,
-        _value = value,
-        _$effect = $effect;
-}
 
 class $UpdateVar<T> extends $Effect {
   final T from;
@@ -90,6 +22,7 @@ class $UpdateVar<T> extends $Effect {
 
 class $AddListener<T> implements $Effect {
   final Function(T) callback;
+  final Type type;
 
   @override
   bool operator ==(other) {
@@ -101,11 +34,12 @@ class $AddListener<T> implements $Effect {
   @override
   int get hashCode => runtimeType.hashCode ^ callback.hashCode;
 
-  $AddListener(this.callback);
+  $AddListener(this.callback) : type = T;
 }
 
 class $RemoveListener<T> implements $Effect {
   final Function(T) callback;
+  final Type type;
 
   @override
   bool operator ==(other) {
@@ -117,7 +51,7 @@ class $RemoveListener<T> implements $Effect {
   @override
   int get hashCode => runtimeType.hashCode ^ callback.hashCode;
 
-  $RemoveListener(this.callback);
+  $RemoveListener(this.callback) : type = T;
 }
 
-final shallowEquals = const IterableEquality().equals;
+class $Ended implements $Effect {}